Default to cstdlib atomics for clang when supported
This makes cstdlib atomics the default for clang if we detect that the
default compilation mode is at least C11 and `__STDC_NO_ATOMICS__` is
not defined.

This is almost identical to #4480, which we originally had to revert
because of some problems with older system headers. Those issues have
been worked around in #12040.