Default to cstdlib atomics for --llvm
This makes cstdlib atomics the default for --llvm. We know the version
of clang/llvm we build will have compiler support for cstdlib atomics
and building clang/llvm requires gcc 4.8 or newer (or a compiler with
equivalent C/C++ features), so we know the system header support will be
there too.