Add warning and test of that warning
This warns the user if they try to delete an opaque array pointer to an array of
arrays.  Our clean up function cannot at the present time fully clean up in that
case, due to not being able to determine the full type of the array from an
arbitrary opaque pointer.

A proper array solution is probably better than trying to fix this leak.