Update based on change in handling of arrays with diff index types
With the advent of opaque array pointers, arrays with index types other than
integers or unsigned integers will now be supported via chpl_opaque_array - this
is because such arrays are recognized as not DefaultRectangular.  Remove the
test that tracks the error message for that situation, and add a comment to the
if statement that generates the error message.  We may expand the support for
chpl_external_array to again include array types that have these unusual indices
and in that case the error message will again trigger and remind us that we need
to think about how to support it (if we haven't already thought about it).