Improve comm=ofi debugging.
These are changes were made in support of diagnosing a bug which led to
"attempt to dereference nil" in a number of tests.  They improve both
reporting and correctness assurance for AMs.  They may have a slight
side effect in non-debug builds, in that the comm bundle for AMs now has
not only the AM operator but also the initiating node in the base part
of the bundle.  This may cause more padding within the bundle and thus
increase its size.  I haven't looked at this hard and so don't know for
sure that it's true, but if so I believe it would only be one additional
byte, so no big deal.

The major positive change in debug mode is that now you can match up
initiated and received AMs by looking for "seqId <initiatorNode>:<seq>"
in the debug output.