Support self-PUT.
There was a sanity check against PUT-from-self in the ofi comm layer,
but it turns out that we actually do that under certain circumstances.
An example is test/optimizations/bulkcomm/asenjo/testGasnet/puts-gets.
In the future it would be nice to catch all the cases of self-PUT
upstream of the comm layers, but for now since this is supported by
comm=gasnet and comm=ugni I'm enabling it in comm=ofi as well.

PR #11962 did the same for self-GETs.