Update string-to-int cast functions to look for a 0x style prefix
In the macro that defines the string-to-int cast functions, check for a prefix
like 0x, 0o, 0b and read the number in the base matching the prefix.

Updated the macro to use strtoll and strtoull instead of strtol so it can
handle 64 bit numbers.  Removed the "bigint" version of the macro that was
previously used for 64 bit numbers.

Removed the .future and .bad from the test asking for this functionality.