Fix support for cstdlib atomics under --llvm (round 2)
In #12045 we started adding `-latomic` to llvm compiles for cstdlib
atomic configurations to try to resolve undefined references to
`atomic_thread_fence`. This ended up being a work around that didn't fix
the core issue, and it didn't work in all cases (e.g. when llvm can't
find a libatomic from gcc >=6.) Here we revert that change and fix this
core issue.

What was actually going wrong was that clang's <stdatomic.h> header has
both a macro and a function prototype for `atomic_thread_fence()` and
our llvm backend isn't able to handle the macro so we ended up using a
function with no implementation, which led to undefined references. New
enough versions of libatomic happen to have an implementation of
`atomic_thread_fence()`, so throwing `-latomic` worked in that case, but
not for older versions of libatomic. To avoid the problem this adds a
`chpl_atomic_thread_fence()` wrapper. Note that we have done similar
things in the past to avoid issues with the llvm backend and macros that
it can't process.