Add test of behavior when sending chpl_opaque_array to argument without dom
I made the choice that array arguments without explicit domains in exported
functions would be treated as chpl_external_arrays.  This test locks in the
behavior when that was not what the user intended (i.e. they probably wanted it
to be a chpl_opaque_array) - the C compiler will tell them they sent the wrong
pointer type.