Rename CHPL_MAKE_SETTINGS_NO_NEWLINES to CHPL_MAKE_CHPLENV_CACHE
We cache our chplenv in Makefile.base so we don't have to run
printchplenv multiple times. We used to store this cache in
`CHPL_MAKE_SETTINGS_NO_NEWLINES`, but this name is internal and pretty
confusing. We set/unset the cache from other places, so rename the
variable to `CHPL_MAKE_CHPLENV_CACHE` to make it more clear what it's
used for.