Remove support for flushing buffered AMOS with non-blocking transactions
For buffered atomics we internally buffer operations, and when flushed
we either use chained transactions (up to 5X performance improvement) or
we had support for issuing non-blocking transactions (up to 2.5X
performance improvement.) This removes the non-blocking transaction
fallback since it's not tested nightly and pre CLE 5.2UP04 is not an
important target for us currently. If it ever becomes important again we
can revert this change, or we could go with something more like
https://github.com/ronawho/chapel-archive/tree/cleanup-nb-amos-ugni,
which refactors code to make the nb-ops part of normal testing paths.
For buffered GETs/PUTs we could use the chpl_comm_nb{put,get} interface.