Change an error around deserialization to use cnames for developers
I was having trouble with deserialization error messages because the
pretty versions of my array types all looked similar in their sugared
form, making it difficult to see what was wrong.  Here, I've changed
the error so that in developer mode, it uses the cnames of the types
rather than their names to aid in comparing the two.