In comm=ofi with fixed threads, provide 1 tx context for non-workers.
If the ofi comm layer finds that it's working with a tasking layer that
uses a fixed number of threads and the ofi provider can supply enough
transmit contexts for each of those to have one of its own, then it will
permanently bind transmit contexts to threads to improve performance.
When checking to see if that was possible we had been ensuring there
were enough contexts for each fixed thread plus one for the AM handler.
But under certain circumstances we may also need one for some other
thread, notably the process itself.  When that happened the context
allocator would find itself out of resources and throw an internal
error.

Here, account for needing one more transmit context, to be shared by all
threads that are neither tasking layer fixed workers nor AM handlers.