Simplify this test
Tracking a compilation error for the library's client program, while
demonstrating that the user will receive an error for sending the wrong type,
led to inconsistent behavior across systems and C compilers.  To avoid having to
juggle all these considerations, just make the test a future tracking a desire
to possibly generate a warning for libraries like this.

Michael did ask me to make this test a future, but since it was so unclear to me
what the intended behavior was beyond what happens today, I originally shied
away from it.  He was right, though, it turns out :p