When using gen_release, if no version set, copy BUILD_VERSION
This allows the commit used to create the release to remain in the
version output. It makes it clearer where releases built solely for
testing come from.

When gen_release is provided with a version argument, the BUILD_VERSION
will not be copied, so it is instead generated to store "0" as is
normally the case for the released version.