Add missing compopts files for --library compilation
I realized belatedly that this test directory didn't have a global COMPOPTS file
to include --library for the .chpl tests, and could due to #11009.  This didn't
impact all the .chpl tests but it did impact one I just added and an older test.
Add a .compopts file which checks the behavior with and without the flag.