Use a reader/writer lock to protect unordered op chains in ugni
Previously, we used a single spinlock to protect any access to the unordered op
linked lists. This meant there could only ever be a single reader or writer.
The buff-atomics-stress test does concurrent ops and flushes to stress the
implementation. Prior to #12113, there was no chain to protect, but adding the
chain meant each flush needs to acquire the list lock. By using a reader/writer
lock we can allow for multiple readers.

With this, the concurrent flushing test is nearly as fast as it was before.