Fix how we prevent GASNet builds from triggering a qthread rebuild
In #12084, we updated the qthreads makefile to call `compileline` with
`CHPL_COMM=none` in order to avoid a false dependency between gasnet and
qthreads. Unfortunately, this caused a different issue where the default value
of `CHPL_TARGET_CPU` differs based on comm=none vs comm!=none. This resulted in
us getting bad include paths to a jemalloc dir that did not exist. Instead of
setting `CHPL_COMM=none` and clearing the chplenv cache, just update the
chplenv cache directly. (replace `CHPL_MAKE_COMM` without touching the rest of
the environment.)