Speed up gasnet builds
We build gasnet for a single conduit/substrate at a time, but we were
leaving gasnet's auto-detect on so it could (and usually did) build
other conduits that we'll never use since we embed the substrate in our
build/install paths.

In order to speed up build times use `--disable-auto-conduit-detect`.
For the cray build we have to explicitly disable conduits that the
cross-configure script enables.