Improvements in network AMO processing.
If we do a direct store to an on-node 'done' indicator following
finishing a 'network' AMO that had to be done using an AM-mediated
processor atomic, do a fence after it.  Otherwise, we don't necessarily
have the ordering guarantees we want.

We were calling the wrong "valid" function for network AMOs.  There are
separate ones for regular, fetching, and comparison AMOs, just as there
are separate functions to initiate those, but in the "vald" check we
were calling the regular one in all cases.  Fix this.

If we have to do a 'network' AMO using an AM-mediated processor atomic
because we don't actually have that particular network AMO capability,
if the target node is our own then just call the processor-atomic
function directly rather than through a self-AMO.