Automatically fence unordered operations at task termination
This adds chpl_comm_task_end as a hook for the comm layer to do things
when a task is ending. Use this to fence any unordered operations.

A more general solution might be to have some way for module code to
register callback functions with  task events like we have in the
runtime, but I'm not sure what we want that to look like or how much
performance overhead it will have. For now just go with hard coding what
is fenced.