Add support for per-task unordered get fences
The previous fence/flush mechanism worked across all threads, which is
pretty expensive. Now that we're looking to be able to fence at task
end, we need a cheaper way to to this. If tasks can't migrate threads,
then we can just flush the current thread. This allows us to have a very
cheap check if there are no pending operations (just a non-atomic
conditional) and if there are pending operations we only have to lock
for the current thread.

If tasks can migrate threads, we have to flush operations for the entire
node still. In our default configuration (flat locale model) tasks
cannot migrate threads, so this flushing should be cheap enough for us
to call it on every task end.