Fix prototype-partial-reduction tests for numa
Some standalone iterators do not work under CHPL_LOCALE_MODEL=numa in
`test/users/engin/partial_reduction_support/modules/dsiMethods.chpl`
because they rely, directly or indirectly, on the standalone these()
in ranges. The latter is not available under NUMA. So under NUMA
the standalone iterators in dsiMethods.chpl are "applicable",
yet do not resolve.

The corresponding resolution failures had been masked by tryToken,
redirecting the compiler to the leader/follower iterators instead,
prior to #12131.

This PR makes the standalone iterators in dsiMethods.chpl not available
when the range standalone iterator is not available, using
`__primitive("method call resolves")`. This is the same solution
that is used in a few other places in the modules and tests.

This PR reinforces the desirability of the following features:

* Enable `Reflection.canResolve()` to accept a mix of param and non-param
actual arguments, as well as be sensitive to named, const, ref actuals etc.
See #10918.

* Language support for iterator forwarding. Lots of iterators
in dsiMethods.chpl redirect to other iterators, with the serial iterator
redirecting to a serial iterator, the leader iterator redirecting
to the corresponding leader iterator, ditto follower and standalone.
As this PR shows, this approach is unnecessarily labor-intensive and
error-prone, esp. in the face of updates to the language/implementation.
See #5255.