Silence a gcc 8 warning about truncated strncat (#12207)
Silence a gcc 8 warning about truncated strncat (#12207)

[reviewed by @ronawho ]

This change adds one byte to a strncat() length limit to take into account the terminating null byte. The existing code was correct, but some installations of gcc 8 complain about string truncation because the terminating null wasn't included. We turned those warnings into errors with -Werror.

So far the problem has only been observed on a Cray XC with CHPL_COMM=ugni and WARNINGS=1.  Tested in that environment and on linux64.