Fix a race with oversubscribed unordered operations
Previously, if there were more tasks than threads we could get into a race
during unordered op initialization. We used to check if the current thread was
initialized, and if not we would grab the initialization lock and then proceed
to initialize. However, trying to acquire the init lock can yield, which meant
we could allow another task on the same thread to run and complete the
initialization, which would cause issues when the original task came back and
redid initialization. To resolve the problem, recheck if initialization was
done after acquiring the init lock.