Update Python files to allow the cleanup function to be called
Add the cleanup function prototype to the chplrt.pxd file.  Moved the generation
of the Python class ChplOpaqueArray to a helper function, and added a new method
to the class.  This new method is called cleanup and will call
cleanupOpaqueArray on the Python class instance's val field.  This method will
need to be called explicitly on the instances (instead of when the object is
cleaned up at the end of the program, due to the library cleanup function not
being called at the end of the program).

I'm going to experiment briefly with supporting calls to this function via
Python's with statement, but I wanted to save a version that works first.