Remove TLS *_buff_thread_info from global list on thread destruction
In ugni when a thread performs a buffered operation for the first time, we
insert its thread local info into a global linked list so we can flush all
threads on a node. When a thread is destroyed that thread local storage is
garbage, so we need to remove it -- add support for that here. In order to do
this we use `pthread_key_create` to create a destructor and insert a dummy call
to `pthread_setspecific` so that the destructor is actually called. The
destructor flushes any pending operations and then removes the thread info from
the linked list.

This doesn't really matter for Chapel tasking layer threads since all current
implementations will keep threads alive until program termination, but it does
matter for potential future tasking layers, and for calls from non tasking
layer threads (say an regular pthread calling into an exported function)