Add support for utilizing the type via a Python with clause
This means that the user will not have to explicitly call the cleanup function
when they're done with the opaque array, they can just obtain it in a with
clause and it will get cleaned up for them when the with clause's scope ends.

I saw this strategy listed in the stackoverflow answer for how to clean up a
Python class and it seemed straight-forward.  Thankfully it was.