Default to auto progress with the sockets provider.
The change to a separate receive context for RMA and AMOs did help, but
some tests still timed out.  However, those tests seemed to run rather
faster if the thread count was greatly reduced (to 2, for example).  My
hypothesis is that the timeouts with more threads were due to contention
and/or cache thrashing related to the pthread mutex used for mutual
exclusion on the rx context completion counter.  It's likely that any
practical solution to this would end up looking something like moving
the progress responsibility onto only one or a few threads.  So here,
when we're using the sockets provider default to automatic progress,
which means the provider manages progress itself, on a separate pthread.
This will maximize communication performance though perhaps slow compute
performance a little in some cases, but when we care about performance
very much we won't be using the sockets provider anyway.