Limit the number of trials for the unordered AMO stress test
This test is timing on on cce+arm after bumping the problem size for the
concurrent flushing in ffbf2f7328. The concurrent flushing puts a lot of
stress on the local spinlocks, which are much slower under cce with
intrinsic processor atomics. Drop the problem size for that config, but
leave it higher when cstdlib atomics are being used.