Simplify how we acquire comm domains in ugni
Don't bother checking the cq_cnt or fork_request buffer before
acquiring a comm domain. It's an extra atomic op and we tend to run
with a thread per core so it's rare to run into cases where the cq_cnt
is high or where the fork_request buffer is already taken.

Basically we now just do a test-and-test-and-set lock to acquire a comm
domain. If we were able to acquire the comm domain, we then check the
cq_cnt and if needed look for a free request buffer. This logic is
easier to follow and I believe will be better for the common case.
With the code reduction, this makes acquire_comm_dom small enough to
inline.

None of these changes seem to impact RA-atomics performance (and it
stresses comm domain acquisition very heavily), but these still seem
like worthwhile cleanups.

If we wanted to we could also make the cq_cnt non-atomic now since we
only access it while the comm domain is acquired. I'm not sure that's
a good idea since that may change in the future and I didn't see any
performance difference from making it non-atomic for ra-atomics.