Changes reflecting review feedback.
Modernize the comment describing the purpose of the second receive
context.

Fix a bug which caused the AM handler's participation in manual progress
to be bursty rather than at intervals, as intended.  Also, add a comment
on that code describing why we do it, since the reason is a bit subtle.