Remove _delete_arr overload and restrict the argument type
Now that eltType is not used in the function itself and instead in a dynamically
dispatched method that is called, the two versions of the function can be
unified into a single one that limits the arr argument to being of type BaseArr.

This allows the chpl_opaque_array wrapper to clean up arrays of arrays.