Make a subclass of BaseArr and use it in all but the rectangular arrays
Defines eltType on the subclass as well as an override of the new method
decEltCountsIfNeeded.  Updates:
- DefaultOpaqueArr,
- BaseSparseArr,
- ArrayViewRankChangeArr,
- DefaultAssociativeArr,
- ArrayViewSliceArr,
- ArrayViewReindexArr,
- UserMapAssocArr, and
- ReplicatedArr
to use the new intervening class.

Tried updating the rectangular ones but backing away from it for now.  If it
proves too difficult, I will add an override to the appropriate one in that
hierarchy (two overrides is better than 9 for sure)