Made array view's serialize / deserialize calls recursive
However, in doing so, we started serializing arrays and domains
for cases where I wasn't expecting to as well (e.g., the 'on e'
clause in chpl_assignDomainWithGetSetIndices()).  This caused
C compilation errors for reasons that weren't clear to me.  To
get things back into a "working" state, I added an extra
argument to the serialize/deserialize routines on Block to prevent
the compiler from calling into it.