Improve stdout+stderr combining.
In nightly testing with python 3, the binary output from mandelbrot-fast
was causing stdout to be bytes instead of a string, and concatenating
that with the (empty) stderr which was still a string wasn't allowed.
Here, rework the stdout+stderr combining to work with either strings or
bytes, and either python 2 or 3.