Fix Python 3 support for functionality added in #12236
We play games with our use of Subprocess Popen/communicate where we
create a wrapper that tries to convert some things from bytes to
strings. We do this because we're trying to maintain python2/3 support
where some tests output byte strings, but generally we need to work with
real strings.

In #12236 we changed how we get the output from communicate(), which
broke the bytes_to_str conversion we were doing. Fix that here by adding
a new helper to merge stdout+stderr in a way that will return the
correct string type. Basically if stdout is a str, then we make sure
stderr is a str. If stdout is a bytes string we convert stderr to a
bytes string if it isn't already a bytes string.

This also fixes support for `useTimedExec=False` (which is never used,
but I was changing code paths around it)