Adjust the mpirun4ofi launcher prediff filtering.
Don't insist that the mpirun messages we filter out start at the
beginning of a line.  At least one test flush()es a partial line and
then does something that causes mpirun to produce one of the filtered
messages.  If we insist on filtering only instances that begin at the
start of a line, we fail to remove this one.