Update lookForBadRTCalls to find calls to atomic_thread_fence
Our runtime atomic fence used to just be `atomic_thread_fence` (same
name as C11 atomics), but in #12100 we changed the name to be
`chpl_atomic_thread_fence`. It's easy to forget about the name change
and it'll happen to work in configurations where cstdlib atomics are the
default. Update our lookForBadRTCalls script to find them.