Make stdout, stderr ordering predictable for most launchers
In #12236 we updated sub_test to not merge stderr into stdout so we
could get predictable ordering for by always having stderr follow stdout
in our testing infrastructure. While this worked for CHPL_LAUNCHER=none
and some launchers it did not work for qsub or slurm based launchers
where stderr was still being merged into stdout.

Here we update chpl_launchcmd and our slurm launchers to split stderr
and stdout. This adds support to the slurm launchers to redirect stderr
to a file. For chpl_launchcmd the major change is to redirect stderr to
a file either with the slurm env var or with manual redirecting for
qsub, the rest of the changes are just piping error_file through the
classes.

Testing:
 - [x] examples+runtime/configMatters on XC slurm w/ and w/o launchcmd
 - [x] examples+runtime/configMatters on CS slurm w/ and w/o launchcmd
 - [x] examples+runtime/configMatters on XC qsub w/ launchcmd
 - [x] examples+runtime/configMatters on XE qsub w/ launchcmd
 - [x] verified we still recognize timeouts on qsub systems