Allow for network AMOs originated from AM handlers.
"Fast" executeOns can do network AMOs.  The comm layer hadn't allowed
for this, because until recently the AM handler ran fast executeOns as
if they were non-fast ones, in tasks.  As such, it didn't have code to
handle the case of waiting for a network AMO completion when the tx
context only had a counter, not a completion queue.  Here, add support
for waiting on a completion counter for a network AMO.  This fixes the
test runtime/configMatters/comm/atomic-on-stack.