Enable parallelism for statement-level promotion
Resolves #11203.

BACKGROUND

Consider a promoted call appearing at statement level, for example:
```chpl
  f(1..5); proc f(i:int){return i;}    // functions/diten/promotedFnValNotCaptured
  Locales.chdir("everyLoc");      // library/..../parallelDirState/chdirAllLocales
```

The call `f(1..5)` resolves to a promoted function, ex. `chpl_promo4_f`,
which is an iterator and so its return type is an `_iteratorRecord`.

Prior to this change, insertReturnTemps() passed the result of this call
to `proc _statementLevelSymbol(ir: _iteratorRecord)`, which ran a serial
for-loop over it. So there was no parallelism there. I could not switch
that to a forall loop because it would be a forall over a formal argument,
which the compiler currently is not equipped to handle.

The call `Locales.chdir("everyLoc")` is different because `chdir` does not
return a value. The promotion implementation handles that case specially
and inserts a forall loop around it. That case is unaffected by this change.
I am working on uniform handling of the two cases in promotion.
Currently the not-returning-a-value case relies on tryToken and I am
working to have it handled like the yes-returning-a-value case.

WITH THIS CHANGE

insertReturnTemps() checks if the statement-level call invokes an iterator
or an iterator forwarder. If so, instead of invoking _statementLevelSymbol(),
it inserts an explicit forall loop (with an empty body) around the call.
This way, the promoted call can execute in parallel when parallelism
is available.

This required running insertDynamicDispatchCalls() AFTER insertReturnTemps().
Otherwise the call here.getChildCount() in the range leader would not be
marked virtual in this test:

    test/functions/diten/promotedFnValNotCaptured.chpl

WHILE THERE

Minor: switch how insertReturnTemps() inserts calls to _statementLevelSymbol()
in the other cases. Now they are inserted relatively to the def of the temp,
for increased clarity.

IF WE NEED TO REVISIT THIS

When looking at the missed dynamic dispatch in promotedFnValNotCaptured.chpl,
I thought that we should handle the statement-level calls right after they are
resolved. Because there are so many things that happen between that and
insertReturnTemps(). I was surprised that only one dynamic dispatch was missed
over our test suite.

So, in an attempt dca139daaa, I moved the statement-level-call processing into
resolveNormalCallFinalChecks(). Turns out that's too early because the return
type of the function is not resolved yet.

So instead in a900d84fda I moved it into resolveExprPhase2(), right after
we resolve the callee. Turns out that there a few other ways that the AST
is different at that point vs. during insertReturnTemps(), which need to
be accounted for:

* The callee may be a param function.
* The call may be the last expression in an ArgSymbol's typeExpr block
  or similar, which should be left alone.
* (The above two cases are hanled in a900d84fda.)
* The call may a recursive call to the enclosing function, in which case
  we do not know whether it should be wrapped in a _statementLevelSymbol().

At this point, I gave up the above attempts and decided just to switch
the order of insertDynamicDispatchCalls() and insertReturnTemps()
and worry about other potential issues as they actually come up.