Fix network atomic FI_CSWAP result handling.
The processor atomic and network atomic interface cmpxchg functions
return boolean true/false to indicate whether or not the cmpxchg matched
the expected value and did the exchange.  But the libfabric network
atomic FI_CSWAP returns the old value of the object as its result, as
all the other libfabric network atomic operations do.  I had neglected
to account for this.  Rework FI_CSWAP handling to synthesize the T/F
return value by comparing the returned result to the expected value.