Prevent overruns in the AM multi-receive buffer.
At least one test (runtime/configMatters/comm/remote-taskspawn-stress)
was failing with a libfabric internal error that turned out to be an
overrun of the inbound AM multi-receive buffer.  Here, set a maximum
message size that is pretty large (10 kb), do a fi_setopt() on the AM
receive context to ensure that our multi-receive buffer always has that
much, and add checks to make sure that we don't send any messages that
are larger than that.  The test passes with this change.

This is not the final work in this area.  We need to be able to do AMs
with arbitrarily large arguments.  But that will require "large call"
support, where the initial AM request is small and the very-large arg
buffer is retrieved from the initiating node by the target node using a
GET.  I'd like to do a full test run with the current code first, to see
where we stand without doing that work.

While here, improve the debug output having to do with AM multi-receive
buffer creation and consumption.