Make mpirun4ofi launcher work with and without a slurm job.
This adjusts the mpirun args emitted by the mpirun4ofi launcher so that
it works whether or not we're within a slurm job.  Outside of slurm it
runs all the program instances on the current node.  Inside of a slurm
job it runs the instances on the allocated nodes in a cyclic manner.