Refine when we can reuse 'd'
This refines when we can reuse 'd' to only do so when its type matches
that of the domain we'd create if we sliced.  But note that this is
now a different definition of slicing than we've traditionally had
because historically, we've intersected 'd' with the array's domain
before taking the slice, permitting a strided array (say) to be sliced
by an unstrided domain.