Add cstdlib-intrinsics to choose atomic implementation as compile time
This adds a cstdlib-intrinsics shim that chooses at compile time whether
to use the intrinsics or cstdlib atomics implementation and makes it the
default for intel and cce.

The shim uses cstdlib atomics when:
 - intel >= 18 with gcc >= 5
 - cce   >= 8.7.7
 - gcc   >= 5.0
 - other compilers that have c11 and don't opt-out of atomic support

Otherwise it falls back to intrinsics.

Only recent versions of intel and cce compilers support cstdlib atomics
but we build our module with older versions. We use atomics for both our
runtime and for the generated code and currently we require that both
use the same implementation (since it's embedded in the runtime path.)
As a result we currently always use intrinsics for cce and intel since
we can't build cstdlib with our gen compilers. To work around that, add
a cstdlib-intrinsics shim that will choose which implementation at
compile time instead of chplenv time. This allows us to build our module
and use intrinsics for the runtime, but still allows user code to use
cstdlib atomics. This also allows users building from source to use
cstdlib-intrinsics and if they're using new enough compilers.

This is a little bit weird, but I think we'll be able to remove this
shim eventually.