Add test reported by user which failed with 1.18 but doesn't today
The test failed in an odd way - when a use/only was present instead of just a
use statement, we would try to look for a writeThis method on qio_regexp_t
(which is an extern type and thus does not have one).  Using a writeln instead
of a writef would also trigger the error.  However, we do not see this behavior
with the current master.