React to learnChapelInYMinutes failure on arm.
A recent change to learnChapelInYMinutes ended up breaking in two ways
in some arm testing, where it hadn't been run before because of a
.skipif.  Here, fix one problem, which was a cut+paste induced typo
introduced about 2 years ago, in an exit() call.  And add code to
preserve the original test output file as a diagnostic aid in the event
the failure seen in nightly testing recurs.  (The .prediff normally
overwrites the original output file.)