Move our python 2/3 compatibility code into a helper module
Move our python 2/3 compatibility shim out of sub_test and into its own
module. This will allow us to use it in other scripts (I want to use it
in chpl_launchcmd.py.) This also  makes it more obvious in sub_test that
something special is happening and we're not just calling the standard
library's subprocess.Popen.