Restore an improved learnChapelInYMinutes.skipif.
In PR 12236 I removed learnChapelInYMinutes.skipif because I believed
that updating sub_test to capture stdout and stderr output separately
and then combine them after the fact would resolve all the variability
in ordering for those two.  Not so.  Although sub_test is fixed there
apparently are launchers which nevertheless still combine stdout and
stderr (see issue #12259), and we can get failures that way (of note:
2019/02/11 Cray XC arm testing with the aprun launcher).

Here, restore the learnChapelInYMinutes.skipif that was there before,
but augment it to allow mpirun4ofi in addition to the other launchers
known to work.

Tested: passes on chapcs with comm=ofi, skipped on XC-arm with
comm=none.