Add memory order argument to network atomic implementation
Previously, we dropped the memory order on the floor in module code. We now
pass it through to the runtime, which still drops it on the floor. In the
future we may be able to optimize based on the memory order, but for now we're
just passing it through so that the compiler can pluck a local  memory order to
use when transforming ordered atomics into unordered atomics.