Allow for address space layout randomization in comm=ofi.
The ofi comm layer didn't work in the presence of address space layout
randomization (ASLR), notably on Mac OS X.  The global and private
broadcast functions assumed that their respective compiler-created
address tables were at the same address in all locales, and for the
private variable tables, that the addresses they contained were the
same across all locales.  Neither of these was actually true, leading
to various sorts of mis-addressing failures (SIGSEGV, SIGBUS, failed
nil checks, etc.).

Here, distribute the necessary addresses among the program instances
during initialization, so that all the nodes have the right addresses
for things they need to reach on other nodes.  As part of this, have
the compiler emit the length of chpl_private_broadcast_table[] into a
new chpl_private_broadcast_table_len variable, so the runtime can know
how long that table is.  comm=gasnet didn't need that because it uses
AMs to do private broadcasts, and comm=ugni didn't need it because
although it does broadcasts using direct PUTs it doesn't have to deal
with ASLR.  But for comm=ofi we want to use direct PUTs even with
ASLR, so we have to know how long the private broadcast table is so we
can share the nodes' copies of that around the job.