Codegen chpl_private_broadcast_table_len properly.
The first cut at this could only emit chpl_private_broadcast_table_len
with the C code generator, not the LLVM one.  And it did so explicitly
rather than using the handy gen-a-global-int function.  Here, solve both
those problems.  The only wrinkle is that chpl_private_broadcast_table
itself is produced in two different places depending on which codegen
we're using, and its length is computed in both of those places.  So we
also gen the _len variable in both, but at least we use nearly identical
calls to do so.