Fix interaction between CHPL_LLVM=system and CHPL_GMP=system on MacOS Mojave (#12352)
Fix interaction between CHPL_LLVM=system and CHPL_GMP=system on MacOS Mojave (#12352)

[reviewed by @mppf ]

Clang will normally search /usr/local/include and /usr/local/lib by default. Some versions of Clang have a bug that causes them to forget to search those directories when a `-isysroot` argument is present. Notably, this happens when upgrading to MacOS Mojave.

The result is that the gmp library installed by Mac homebrew will not be found. This change works around the bug by adding the missing directories at the appropriate place in the search path when a `-isysroot` argument is seen.

Tested on MacOS Sierra (10.12) and Mojave (10.14) as well as linux64.