Implement zippered serial forall loops
Handle the case of a forall loop, represented as a ForallStmt, that is:

* zippered,
* "serial OK", i.e. accepts serial iterators,
* invokes serial iterators because parallel iterators are not available.

The loop may be either explicit as a `[]` statement or a `[]` expression
or implicit in a reduce expression or a promoted expression.

Such a loop is to be executed as a zippered **serial** loop, so we convert it
to a ForLoop. We keep the original ForallStmt as a wrapper around this ForLoop
by giving it an iterator that yields a single, unused value.
Retaining the ForallStmt itself means that forall intents, if any,
are handled by existing code.

Supporting change: have resolveForallHeader() distinguish explicitly
these two cases:

* The main iterator is standalone, i.e. obtained by adding
  `tag=iterKind.standalone` to the iterable call.

* The main iterator observed at start of resolveForallHeader()
  is used as-is.

This latter case occurs for example when the ForallStmt is to iterate
over the serial iterator or when the ForallStmt is the result of converting
a ForLoop iterating over a parallel iterator.

Resolves #12196.